+++ src/currency/currency.entity.ts
@@ -1,0 +1,4 @@
+export class Currency {
+  currency: string;
+  value: number;
+}
