+++ src/currency/currency.module.ts
@@ -1,5 +1,6 @@
 import { Module } from '@nestjs/common';
-import { CurrencyRepository, CurrencyService } from './currency.service';
+import { CurrencyRepository } from './currency.repository';
+import { CurrencyService } from './currency.service';
 
 @Module({
   providers: [CurrencyService, CurrencyRepository]
